requirements.txt
streamlit>=1.35.0
pypdf2>=3.0.1
chromadb>=0.5.0
sentence-transformers>=2.7.0
ollama>=0.1.6
python-dotenv>=1.0.1
fpdf2
pandas